Made changes to my user database functions, the side drawer, and the edit location and username files.
The side drawer now uses a stream builder instead of a future builder.  Ideally, this means that less queries will be used in return for having a listener that continues running in the background.  The side drawers were removed from the edit location and edit username in order to prevent the user from creating a cycle of pages where they hop from side drawer to edit page back to another side drawer.