Implemented the edit username page.
The edit username page now allows a user to update their username while preventing them from using a username already in use by another user.  I added a basic utility class to act as a wrapper for some common mediaquery functions that give us a size based on the size of the screen.  Our models were changed to align with the fact that our database function now uses strings for the barcode.