Added store information to the sidebar so that it can be passed to the edit_location page.
I had to change the user class to represent zip codes as strings since we are now storing them that win in our database.  The store class has more helper methods to make it easier to access different parts of the address quickly.  The store class now also has a constructor that will build a store item from a map.  This is mostly used for the maps I receive from the database.