Changed edit_location and side_drawer to take advantage of our change to the database.
Our database now stores the user's preferred location as a map containing all the information associated with that store.  It should be a fairly low cost change since it simply sends more data with the one update request when their location is changed.  This allows us to have easy access to the user's preferred store data without needing to make a second request for the store's information.  An immediate benefit is shown in the changes to the side_drawer that allowed me to remove the extra code and validation needed when making a query.