The shopping list is pulling all items from a user's shopping list and allowing them to change the quantity while the item is still in their list.
The quantity correctly updates in the database when changed.  No query is made when the same quantity is selected.  The current formatting appears to work in both landscape and portrait mode, but the image in landscape mode could be better.  The list item class was added to help represent the information found within shopping list items.  While I liked the idea of using inheritance within all the items, I've increasingly found it seems to be more confusing than helpful.