Changed the structure of the app to accommodate the shopping list's removal from the bottom navigation bar.  Includes minor changes to the models in order to align them with their current use.
We decided to have the shopping list in the top right like most shopping apps seem to prefer.  This should be a familiar pattern to users.  When in the shopping list, the previous shopping list icon will be replaced with the home icon to take them back to their previous page.